Made files little more organized and added Camera class

+++ Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using OpenTK.Graphics.OpenGL4;
 
+using opentk_proj.util;
+
 namespace opentk_proj
 {
     internal class Program
@@ -8,7 +10,7 @@ namespace opentk_proj
         static void Main(string[] args)
         {
 
-            using (Game game = new Game((int)Constants.WIDTH, (int)Constants.HEIGHT, "Game"))
+            using (Game game = new Game(Constants.WIDTH, Constants.HEIGHT, "Game"))
             {
                 game.Run();
